gateway/requirements.txt
# Gateway dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
asyncpg>=0.29.0
PyJWT>=2.8.0
bcrypt>=4.1.0
pyyaml>=6.0.0
structlog>=24.0.0
email-validator>=2.1.0
